fix multiple bugs on sp layer
